# docs/review-api-openapi.yaml
openapi: 3.0.3
info:
  title: Review API
  description: API untuk mengelola ulasan produk dengan fitur CRUD lengkap, statistik, dan manajemen admin
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://your-domain.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /reviews/stats:
    get:
      summary: Get Review Statistics
      description: Mengambil statistik ulasan untuk produk atau keseluruhan
      tags:
        - Statistics
      parameters:
        - name: productId
          in: query
          description: Statistik untuk produk tertentu
          required: false
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewStatsResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Review:
      type: object
      properties:
        id:
          type: integer
          example: 1
        productId:
          type: integer
          example: 1
        userId:
          type: string
          format: uuid
          example: "a4bc7b55-bee9-4f13-8486-9cb8bb92be29"
        userName:
          type: string
          example: "John Doe"
        userAvatar:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Produk sangat bagus!"
        date:
          type: string
          format: date
          example: "2024-01-15"
        verified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Rating harus antara 1-5"
    RatingDistribution:
      type: object
      properties:
        "1":
          type: integer
          example: 1
        "2":
          type: integer
          example: 2
        "3":
          type: integer
          example: 4
        "4":
          type: integer
          example: 8
        "5":
          type: integer
          example: 10
    TopRatedProduct:
      type: object
      properties:
        productId:
          type: integer
          example: 1
        productName:
          type: string
          example: "Laptop Gaming ASUS"
        averageRating:
          type: number
          format: float
          example: 4.8
        totalReviews:
          type: integer
          example: 50
    ReviewStatsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            productId:
              type: integer
              example: 1
            totalReviews:
              type: integer
              example: 25
            averageRating:
              type: number
              format: float
              example: 4.2
            ratingDistribution:
              $ref: '#/components/schemas/RatingDistribution'
            verifiedReviews:
              type: integer
              example: 20
            unverifiedReviews:
              type: integer
              example: 5
            recentReviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            topRatedProducts:
              type: array
              items:
                $ref: '#/components/schemas/TopRatedProduct'
...